// modules/redux/order/order.enum.ts
export enum OrderActionTypes {
  ORDER_ADD_CLIENT_INFORMATION_START = 'ORDER_ADD_CLIENT_INFORMATION_START',
  ORDER_ADD_CLIENT_INFORMATION_SUCCESS = 'ORDER_ADD_CLIENT_INFORMATION_SUCCESS',
  ORDER_ADD_CLIENT_INFORMATION_FAILURE = 'ORDER_ADD_CLIENT_INFORMATION_FAILURE',
  ORDER_CREATE_START = 'ORDER_CREATE_START',
  ORDER_CREATE_SUCCESS = 'ORDER_CREATE_SUCCESS',
  ORDER_CREATE_FAILURE = 'ORDER_CREATE_FAILURE',
  ORDER_ADD_DELIVERY_FEE = 'ORDER_ADD_DELIVERY_FEE',
}

export enum OrderSelectorEnum {
  Loading = 'loading',
  Order = 'order',
  Error = 'error',
  DeliveryFee = 'deliveryFee',
  Client = 'client',
  Data = 'data',
}